requirements.txt
requests==2.31.0
dbus-python==1.3.2
pyyaml==6.0
schedule==1.2.0 